feat: add participation-based exp bonus system
Refactor SessionData.exp from long to float to support bonus calculations
Add expGainBonus float field to Session class
Implement participation-based exp bonus in SurvivalRank, calculating based on player's map play time and applying to exp gains
Fix RtvService to fire GameOverEvent for Team.neoplastic instead of Team.crux
Update all usages of SessionData.exp to cast float to long where required by existing methods

+++ plugin/src/main/java/plugin/session/ExpUtils.java
@@ -3,7 +3,7 @@ package plugin.session;
 public class ExpUtils {
 
     public static long getLevel(Session session) {
-        return levelFromTotalExp(session.getData().exp);
+        return levelFromTotalExp((long) session.getData().exp);
     }
 
     // 1 EXP per second of play time
