+++ plugin/src/main/java/plugin/session/SessionData.java
@@ -6,7 +6,7 @@ public class SessionData {
     // in ms
     public String name = new String();
     public long playTime = 0;
-    public long exp = 0;
+    public float exp = 0;
     public long lastSaved = Instant.now().toEpochMilli();
     public String trail = new String();
     public boolean griefer = false;
